+++ src/security_log/aiwaf.rs
@@ -1,11 +1,14 @@
 use std::{net::IpAddr, str::FromStr, sync::OnceLock};
 
 use anyhow::{anyhow, bail, Context, Result};
-use chrono::{DateTime, FixedOffset};
 use giganto_client::ingest::log::SecuLog;
+use jiff::Timestamp;
 use regex::Regex;
 
-use super::{Aiwaf, ParseSecurityLog, SecurityLogInfo, DEFAULT_IPADDR, DEFAULT_PORT, PROTO_TCP};
+use super::{
+    timestamp_to_i64, Aiwaf, ParseSecurityLog, SecurityLogInfo, DEFAULT_IPADDR, DEFAULT_PORT,
+    PROTO_TCP,
+};
 
 fn get_aiwaf_regex() -> &'static Regex {
     static LOG_REGEX: OnceLock<Regex> = OnceLock::new();
@@ -16,8 +19,8 @@ fn get_aiwaf_regex() -> &'static Regex {
     })
 }
 
-fn parse_aiwaf_timestamp(datetime: &str) -> Result<DateTime<FixedOffset>> {
-    DateTime::parse_from_str(&format!("{datetime} +0900"), "%Y-%m-%d %H:%M:%S %z")
+fn parse_aiwaf_timestamp(datetime: &str) -> Result<Timestamp> {
+    Timestamp::strptime("%Y-%m-%d %H:%M:%S %z", format!("{datetime} +0900"))
         .map_err(|e| anyhow!("{e:?}"))
 }
 
@@ -56,10 +59,7 @@ impl ParseSecurityLog for Aiwaf {
             None => DEFAULT_PORT,
         };
 
-        let timestamp = parse_aiwaf_timestamp(datetime)?
-            .timestamp_nanos_opt()
-            .context("to_timestamp_nanos")?
-            + serial;
+        let timestamp = parse_aiwaf_timestamp(datetime).and_then(timestamp_to_i64)? + serial;
 
         Ok((
             SecuLog {